+++ src/layer/interface/module/view/scroll.ts
@@ -11,7 +11,7 @@ export class ScrollView extends Coroutine<never> {
   ) {
     super(async function* () {
       return this.finally(bind(window, 'scroll', debounce(100, ev => {
-        if (standardize(window.location.href) !== page.href) return;
+        if (standardize(window.location.href) !== page.url.href) return;
         void listener(ev);
       }), { passive: true }));
     }, { delay: false });