+++ src/layer/interface/service/api.ts
@@ -1,6 +1,7 @@
 import { Config as Option } from '../../../../';
-import { route, sync, Config, RouterEvent, RouterEventSource } from './router';
+import { route, Config, RouterEvent, RouterEventSource } from './router';
 import { process } from './state/process';
+import { page } from './state/page';
 import { savePjax, isTransitable } from '../../data/store/state';
 import { parse } from '../../../lib/html';
 import { assign } from 'spica/assign';
@@ -10,21 +11,21 @@ export class API {
   public static assign(url: string, option: Option, io = { document: window.document, router: route }): boolean {
     let result!: boolean;
     void click(url, event =>
-      result = io.router(new Config(option), new RouterEvent(event), process, io));
+      result = io.router(new Config(option), new RouterEvent(event, page.url), process, io));
     assert(result !== void 0);
     return result;
   }
   public static replace(url: string, option: Option, io = { document: window.document, router: route }): boolean {
     let result!: boolean;
     void click(url, event =>
-      result = io.router(new Config(assign({}, option, { replace: '*' })), new RouterEvent(event), process, io));
+      result = io.router(new Config(assign({}, option, { replace: '*' })), new RouterEvent(event, page.url), process, io));
     assert(result !== void 0);
     return result;
   }
   public static sync(isPjaxPage?: boolean): void {
     isPjaxPage && void savePjax();
     void process.cast('', new Error(`Canceled.`));
-    void sync();
+    void page.sync();
   }
   public static pushURL(url: string, title: string, state: unknown = null): void {
     void window.history.pushState(state, title, url);