+++ src/layer/interface/service/router.test.ts
@@ -1,4 +1,5 @@
 import { _validate as validate, Config, RouterEvent } from './router';
+import { page } from './state/page';
 import { parse } from '../../../lib/html';
 import { URL, standardize } from 'spica/url';
 import { html, delegate, once } from 'typed-dom';
@@ -9,7 +10,7 @@ describe('Unit: layer/interface/service/router', function () {
       const url = '';
       const document = parse(`<a href="${url}"></a>`).extract();
       delegate(document, 'a', 'click', ev => {
-        assert(validate(new URL(standardize(url, window.location.href)), new Config({}), new RouterEvent(ev)));
+        assert(validate(new URL(standardize(url, window.location.href)), new Config({}), new RouterEvent(ev, page.url)));
         done();
       });
       document.querySelector('a')!.click();
@@ -19,7 +20,7 @@ describe('Unit: layer/interface/service/router', function () {
       const url = '//external';
       const document = parse(`<a href="${url}"></a>`).extract();
       delegate(document, 'a', 'click', ev => {
-        assert(!validate(new URL(standardize(url, window.location.href)), new Config({}), new RouterEvent(ev)));
+        assert(!validate(new URL(standardize(url, window.location.href)), new Config({}), new RouterEvent(ev, page.url)));
         ev.preventDefault();
         done();
       });
@@ -30,7 +31,7 @@ describe('Unit: layer/interface/service/router', function () {
       const url = '#';
       const document = parse(`<a href="${url}"></a>`).extract();
       delegate(document, 'a', 'click', ev => {
-        assert(!validate(new URL(standardize(url, window.location.href)), new Config({}), new RouterEvent(ev)));
+        assert(!validate(new URL(standardize(url, window.location.href)), new Config({}), new RouterEvent(ev, page.url)));
         ev.preventDefault();
         done();
       });
@@ -41,7 +42,7 @@ describe('Unit: layer/interface/service/router', function () {
       const url = '';
       const document = parse(`<a href="${url}" download></a>`).extract();
       delegate(document, 'a', 'click', ev => {
-        assert(!validate(new URL(standardize(url, window.location.href)), new Config({}), new RouterEvent(ev)));
+        assert(!validate(new URL(standardize(url, window.location.href)), new Config({}), new RouterEvent(ev, page.url)));
         ev.preventDefault();
         done();
       });
@@ -55,7 +56,7 @@ describe('Unit: layer/interface/service/router', function () {
       ]);
       document.body.appendChild(form);
       once(document, 'form', 'submit', ev => {
-        assert(validate(new URL(standardize(url, window.location.href)), new Config({}), new RouterEvent(ev)));
+        assert(validate(new URL(standardize(url, window.location.href)), new Config({}), new RouterEvent(ev, page.url)));
         ev.preventDefault();
         form.remove();
         done();
@@ -70,7 +71,7 @@ describe('Unit: layer/interface/service/router', function () {
       ]);
       document.body.appendChild(form);
       once(document, 'form', 'submit', ev => {
-        assert(!validate(new URL(standardize(url, window.location.href)), new Config({}), new RouterEvent(ev)));
+        assert(!validate(new URL(standardize(url, window.location.href)), new Config({}), new RouterEvent(ev, page.url)));
         ev.preventDefault();
         form.remove();
         done();