+++ src/layer/interface/service/state/script.ts
@@ -1,3 +1,4 @@
+import { page } from './page';
 import { URL, StandardURL, standardize } from 'spica/url';
 import { bind } from 'typed-dom/listener';
 
@@ -6,4 +7,4 @@ export const scripts = new Set<URL.Href<StandardURL>>();
 void bind(window, 'pjax:unload', () =>
   void document.querySelectorAll('script[src]')
     .forEach(script =>
-      void scripts.add(new URL(standardize(script.src)).href)));
+      void scripts.add(new URL(standardize(script.src, page.url.href)).href)));